[UI] Move the gate test out of the Next.js pages directory
ui/next.config.js sets no pageExtensions, so Next's default
['tsx','ts','jsx','js'] applies to every file under ui/pages/ - a
previously generated ui/.next/required-server-files.json records exactly
that resolved value. A test placed there is therefore compiled as the
route /configuration/designs/configurator.test, has no default export,
and fails `next build` - which CI runs via make ui-build.

Move it to ui/__tests__/pages/configuration/designs/, matching the
existing __tests__/architecture/ precedent. vitest's include already
covers **/__tests__/**, so the deny-path coverage is unchanged.

Adding a pageExtensions setting was rejected: it would change route
resolution for the whole app to work around one misplaced file.

Also corrects the entry-point comment's grouping - buildPatternColumns
never receives arePatternsReadOnly and ViewSwitch renders
unconditionally, so the designs-list row action is reachable from the
catalog page's table view too.
