+++ ui/pages/configuration/designs/configurator.tsx
@@ -10,13 +10,20 @@ import { MesheryPage } from '@/components/general/MesheryPage';
  *
  * Four call sites navigate here:
  *
- * 1. `components/designs/patterns/MesheryPatterns.tsx:365` - the designs list
- *    row action.
- * 2. `components/designs/patterns/MesheryPatternsToolbar.tsx:61` - the create
- *    new design toolbar button.
+ * 1. `components/designs/patterns/MesheryPatternsToolbar.tsx:61` - the create
+ *    new design toolbar button. This is the only one exclusive to the designs
+ *    page.
+ * 2. `components/designs/patterns/MesheryPatterns.tsx:365` -
+ *    `handleOpenInConfigurator`, behind the table view's "Edit" row actions
+ *    (`MesheryPatterns.columns.tsx:50` and `:72`). It is NOT exclusive to the
+ *    designs page: `buildPatternColumns` is never passed `arePatternsReadOnly`
+ *    and the toolbar's `ViewSwitch` renders whenever no design detail is open,
+ *    so the same action is reachable in table view on `/configuration/catalog`.
+ *    Both row actions key on `CatalogManagementEditDesign`, not ViewDesigns.
  * 3. `components/designs/patterns/MesheryPatternCard.tsx:103` - "Edit In
- *    Configurator" on the design card, reachable from `/configuration/catalog`,
- *    which gates on `CatalogManagementViewCatalog` rather than ViewDesigns.
+ *    Configurator" on the design card, the grid-view counterpart of the above
+ *    and likewise reachable from `/configuration/catalog`, which gates on
+ *    `CatalogManagementViewCatalog` rather than ViewDesigns.
  *    `arePatternsReadOnly` does not suppress that button: the `isReadOnly`
  *    guard at `MesheryPatternCard.tsx:441` covers a different block, so the
  *    button renders whenever `userCanEdit` (holds `CatalogManagementEditDesign`,
@@ -25,23 +32,30 @@ import { MesheryPage } from '@/components/general/MesheryPage';
  *    workspace design open, which falls back to the configurator when the
  *    Kanvas designer is unavailable.
  *
- * `CatalogManagementViewDesigns` is the same key that gates the first two: the
- * designs list (`MesheryPatterns.tsx:72`) and the Designs navigator entry
- * (`components/layout/Navigator/navigatorComponents.tsx:159`). It remains the
- * single correct key despite the last two, because no built-in keychain in
- * `server/permissions/keys.csv` holds `Edit Design`, `View Catalog` or
- * `View Workspace` without also holding `View Designs` - verified per role
+ * So three of the four can originate from a page that is not gated on
+ * ViewDesigns - the catalog card button, the catalog table row action, and the
+ * workspace Kanvas fallback.
+ *
+ * `CatalogManagementViewDesigns` is nonetheless the correct single key. It is
+ * what already gates the designs list (`MesheryPatterns.tsx:72`) and the
+ * Designs navigator entry
+ * (`components/layout/Navigator/navigatorComponents.tsx:159`), and no built-in
+ * keychain in `server/permissions/keys.csv` holds `Edit Design`, `View Catalog`
+ * or `View Workspace` without also holding `View Designs` - verified per role
  * column across User, Team Admin, Academy Admin, Leaner, Workspace Admin, Org
  * Billing Manager, Org Admin and Provider Admin. No default role can reach a
- * deny page from those buttons; only a custom keychain can.
+ * deny page from those three; only a custom keychain can.
  *
  * Where a custom keychain does make that dead end reachable, the fix belongs at
- * the entry-point buttons - a button that leads to a permission-denied page
- * should not render - and not in widening this gate. The configurator reads and
- * writes full design JSON, whereas `CatalogManagementViewCatalog` grants only
- * read-only published browsing, so accepting it here would grant more than the
- * designs list itself does. That button-side fix is deliberately out of scope
- * here and raised as a separate product question.
+ * the entry points - a control that leads to a permission-denied page should
+ * not render - and not in widening this gate. Note that this means fixing both
+ * spellings: `MesheryPatternCard.tsx` for grid view AND
+ * `MesheryPatterns.columns.tsx` for table view, or the table path still dead
+ * ends. The configurator reads and writes full design JSON, whereas
+ * `CatalogManagementViewCatalog` grants only read-only published browsing, so
+ * accepting it here would grant more than the designs list itself does. That
+ * entry-point fix is deliberately out of scope here and raised as a separate
+ * product question.
  *
  * Gating here rather than inside `DesignConfigurator` keeps the component from
  * mounting at all, so the unconditional `useMeshModelComponents` category fetch
